+++ changelog.md
@@ -17,6 +17,7 @@
 * Zombie Villagers without any family won't be buried
 * Infected villagers being killed by a zombie no longer duplicate their inventory
 * Infection now lasts longer
+* Adventurers with high hearts may stay without asking
 
 # 7.3.14